requirements.txt
patool==1.12
psycopg2-binary==2.9.3
rarfile==4.0